requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
openai==1.58.1
google-genai>=1.51.0
pyyaml==6.0.2
pydantic-settings==2.7.1
